refactor: move the preflight analysis core into @lightdash/common
The analysis lived in scripts/preflight.ts, which a published package
cannot import, so the CLI command threw at a seam rather than running.
The pure core now lives in packages/common/src/preflight and both
consumers import it from there.

scripts/preflight.ts keeps its IO shell and re-exports the core, so its
existing tests and gen-migration-facts.ts need no changes — the point of
this shape is that the repo's release-safety gate keeps testing exactly
what it tested before. expand-version and json-schema-lite are moved and
re-exported the same way.

parseFactsFile read its schema from disk, which a published package
cannot rely on, so the schema moves into common and is imported
directly, following the JSON schemas already bundled there.

Deriving the operator's current version stays unwired and --from stays
required: that needs the probe endpoint to return applied migrations,
which lands separately.

Relates: SPK-894
